+++ lang/examples/advanced/ffi_http/http_client/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "http_client"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+reqwest = { version = "0.12.11", features = ["blocking"] }
+
+[lib]
+crate-type = ["cdylib"]
