+++ lang/examples/advanced/ffi_http/main.modu
@@ -1,0 +1,28 @@
+import "os" as os;
+import "ffi" as ffi;
+
+fn check_comp() {
+    if os.name != "linux" {
+        print("libhttp only supports linux at the current moment");
+        exit();
+    }
+}
+
+fn get(url) {
+    let res = ffi.call("./libhttp_client.so", "get", url);
+
+    return res;
+}
+
+check_comp();
+print(get("https://jsonplaceholder.typicode.com/todos/1"));
+
+// Expected output:
+/*
+{
+    userId: 1,
+    id: 1,
+    title: delectus aut autem,
+    completed: false
+}
+*/
