--- lang/other/packages/libhttp/http_client/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "http_client"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-reqwest = { version = "0.12.12", features = ["blocking"] }
-
-[lib]
-crate-type = ["cdylib"]
