--- lang/other/packages/libhttp/lib.modu
@@ -1,19 +1,0 @@
-import "os" as os;
-import "ffi" as ffi;
-
-fn check_comp() {
-    if os.name == "linux" {
-        return true;
-    }
-
-    print("Only linux is supported at the current moment");
-    exit();
-}
-
-fn get(url) {
-    let res = ffi.call("./libhttp_client.so", "get", url);
-
-    return res;
-}
-
-print(get("http://google.com"));
