--- lang/other/packages/libhttp/project.toml
@@ -1,3 +1,0 @@
-[package]
-name = "libhttp"
-version = "1.0.0"
